Better reflect current SOP with not activated ID
This PR changes the message to the user what he should do when he is using a not online-activated ID.
Example: An ID is issued and the online function is not activated. The responsible regsitry office (Einwohnermeldeamt) is called A.
The user moves and another is now another registry office (B) is responsible for the user. The user wishes to user AusweisApp2 to authenticate.
In the current state the message can be understood to contact registry office A, which in this case is not responsible for the user.