CMake: Declare platform libraries to be explicitly static.
The libraries declared in CMake using the custom `ADD_PLATFORM_LIBRARY`
function are intended for internal consumption by AusweisApp2.  These
libraries do NOT have a proper so-name, neither get properly installed
by CMake, nor header files for integration into external applications
are provided.

In most of the rpm-based Linux distributions the macroized invocation
of CMake assumes any non-explicitly as STATIC or OBJECT declared
library to be implicitly built as SHARED.

Thus libraries that are intended to be a 'platform' library for the
AusweisApp2 application need to be explicitly declared as STATIC.

Signed-off-by: Björn Esser <besser82@fedoraproject.org>