Doxyfile: Exclude binary, CMake, libs, and test directories.
These directories do not contain any useful source files
for the documentation of the internal API of AusweisApp2,
and thus should not be included in its API documentation
generated by Doxygen.

Additionally this comes in handy for distributing the API
documentation architecture independent, when the name of
the binary directory contains the name of the system's
architecture the build is targeted to.

Also explicitly set the 'FULL_PATH_NAMES' parameter to 'YES',
as this is needed for properly stripping and/or excluding
the paths during generation of the documentation files.

Signed-off-by: Björn Esser <besser82@fedoraproject.org>